reflect, internal/reflectlite: embed reflectlite types into reflect types
